refactor downloader code
