+++ crates/common/download/src/lib.rs
@@ -1,5 +1,7 @@
 mod download;
 mod error;
 
+pub use crate::download::Auth;
+pub use crate::download::DownloadInfo;
 pub use crate::download::Downloader;
 pub use crate::error::DownloadError;
